Add event_kind to message instead of dropping it after using it for filtering. verify by inspection of trace output to std::cout sink

+++ src/libraries/tracing/include/m/tracing/event_context.h
@@ -61,8 +61,8 @@ namespace m
 
         protected:
             std::thread::id m_thread_id;
-            uint64_t        m_os_process_id;
-            uint64_t        m_os_thread_id;
+            uint32_t        m_os_process_id;
+            uint32_t        m_os_thread_id;
             utc_time_point  m_time_point;
         };
